refactor constants
