+++ dataSources/parquet/src/main/java/cn/edu/tsinghua/iginx/parquet/common/Constants.java
@@ -1,0 +1,20 @@
+package cn.edu.tsinghua.iginx.parquet.common;
+
+public final class Constants {
+
+  public static final String IGINX_SEPARATOR = "\\.";
+
+  public static final String PARQUET_SEPARATOR = "\\$";
+
+  public static final String SUFFIX_PARQUET_FILE = ".parquet";
+
+  public static final String SUFFIX_EXTRA_FILE = ".extra";
+
+  public static final String CMD_PATHS = "PATHS";
+
+  public static final String CMD_KEY = "TIME";
+
+  public static final String CMD_DELETE = "DELETE";
+
+  public static final int MAX_MEM_SIZE = 1000 * 1024 * 1024 /* BYTE */;
+}
